Tommy Holmes Removed the Primary and Foreign key creation as those are located in separate scripts.  I also adjusted ordering of columns and alphabetized table names.
I also made some major changes that I'm not sure of:

Removed column 'buildingFeeID' as it was not created as a foreign key and nothing references it except the trigger which also references REVENUETYPE (so it's out of date).  Not sure we even need BILLINGDETAIL anymore considering it looks like PAYMENTBILLING handles the same thing.

Removed column 'utilityTypeID' from table 'BULIDINGFEETYPE' as it was not created as a foreign key.  It is referenced in the trigger 'BuildingFeeTypeInsertUpdate'.

Removed column 'documentPermissionLevel' in 'DOCUMENT' as it seems to do the same thing as 'documentIsDisabled' in that they can restrict who sees the document by that.

Removed column 'buildingID' in 'EMPLOYEETRANSACTION' as '[TRANSACTION]' requires a 'buildingID' and 'EMPLOYEETRANSACTION' requires a 'transactionID'.